Binding data to layout

+++ WPF.Client/Models/RPCModel.cs
@@ -6,7 +6,7 @@
         public int Port { get; set; }
         public string Image { get; set; }
         public int TransferDataSize { get; set; }
-        public int TotalExecutionTime { get; set; }
+        public long TotalExecutionTime { get; set; }
         public string Type { get; set; }
     }
 }
